chore(clerk-js): Cleanup build files
