+++ packages/clerk-js/package.json
@@ -26,9 +26,9 @@
     "build": "npm run build:bundle && (npm run build:declarations || true) && npm run bundlewatch",
     "build:stats": "npm run build --json > stats.json",
     "bundlewatch": "bundlewatch",
-    "build:bundle": "webpack --config webpack.dev.v4.js --env production",
+    "build:bundle": "webpack --config webpack.dev.js --env production",
     "build:declarations": "tsc -p tsconfig.declarations.json",
-    "dev": "webpack serve --config webpack.dev.v4.js",
+    "dev": "webpack serve --config webpack.dev.js",
     "prepublishOnly": "npm run build",
     "postpublish": "node ./scripts/purge-cache.mjs",
     "start": "echo \"Noop\"",
