--- packages/clerk-js/tsconfig.stable.json
@@ -1,25 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": false,
-    "baseUrl": "src",
-    "declaration": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "module": "esnext",
-    "moduleResolution": "Node",
-    "noEmit": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitReturns": true,
-    "noUnusedLocals": false,
-    "noUnusedParameters": true,
-    "resolveJsonModule": true,
-    "sourceMap": true,
-    "strict": true,
-    "target": "ES2019",
-    "useUnknownInCatchVariables": false
-  },
-  "exclude": ["node_modules"],
-  "include": ["src/**/*"]
-}
