--- packages/clerk-js/tsconfig.v4.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.stable.json",
-  "compilerOptions": {
-    "declaration": false,
-    "jsx": "react-jsx",
-    "jsxImportSource": "@emotion/react"
-  },
-  "exclude": ["node_modules", "**/*.test.ts", "**/*.test.tsx"],
-  "include": ["src/**/*"]
-}
